Added the 'REVISED_C971_Wireframe.fig' file to the Resources/Wireframe folder after making adjustments. Left the 'OLD_NOT_CURRENT' version in the folder for visbility of change.
I strayed from the original wireframe because I quickly realized that having only one page to add/edit/delete terms, assessments, and courses was becoming very confusing and messy. Pushing and Poping was a much better way to go about navigating. The UI looks much different as well because as I was becoming more familiar with using XAML, I knew more what I had to work with.

Lastly, The overall logic and presentation of the finished application and wireframe is much more robust.